expose Logger interface
